+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             android:name=".WriteNotesActivity"
             android:label="@string/new_note"></activity>
         <activity
-            android:name=".NotesAct"
+            android:name=".NotesListActivity"
             android:label="@string/title_activity_notes"></activity>
         <activity android:name=".MainActivity">
             <intent-filter>
