--- app/src/main/java/com/example/notesandreminding/NotesAct.java
@@ -1,118 +1,0 @@
-package com.example.notesandreminding;
-
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
-
-import java.util.ArrayList;
-
-public class NotesAct extends AppCompatActivity {
-    private ListView listView;
-    private MapAdapter adapter;
-
-    @Override
-    public void onSaveInstanceState(Bundle savedInstanceState) {
-        super.onSaveInstanceState(savedInstanceState);
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_notes);
-        listView = findViewById(R.id.listView);
-
-        findViewById(R.id.fab).setOnClickListener(addNoteClickListener);
-        listView.setOnItemClickListener(itemClickListener);
-        listView.setOnItemLongClickListener(onItemLongClickListener);
-
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        showNotes(App.getNoteRepository().getNotes());
-    }
-
-    AdapterView.OnItemLongClickListener onItemLongClickListener = new AdapterView.OnItemLongClickListener() {
-        @Override
-        public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-
-            final Note val = adapter.mData.get(position);
-
-            AlertDialog.Builder builder = new AlertDialog.Builder(NotesAct.this)
-                    .setTitle(R.string.warning)
-                    .setMessage(R.string.Sure)
-                    .setCancelable(false)
-                    .setPositiveButton(R.string.delete, new DialogInterface.OnClickListener() {
-                        public void onClick(DialogInterface dialog, int arg1) {
-                            App.getNoteRepository().removeNote(val.getId());
-                            showNotes(App.getNoteRepository().getNotes());
-                        }
-                    });
-            builder.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                public void onClick(DialogInterface dialog, int arg1) {
-                    dialog.cancel();
-                }
-            });
-
-            AlertDialog alert = builder.create();
-            alert.show();
-            return true;
-        }
-    };
-
-    AdapterView.OnItemClickListener itemClickListener = new AdapterView.OnItemClickListener() {
-        @Override
-        public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-            final Note val = adapter.mData.get(position);
-            final long ID = val.getId();
-            if (val != null) {
-                Intent intent = new Intent(NotesAct.this, WriteNotesActivity.class);
-                intent.putExtra("ID", ID);
-                startActivity(intent);
-            }
-        }
-    };
-
-    private void showNotes(ArrayList<Note> not) {
-        adapter = new MapAdapter(not);
-        listView.setAdapter(adapter);
-    }
-
-    private View.OnClickListener addNoteClickListener = new View.OnClickListener() {
-
-        @Override
-        public void onClick(View v) {
-            finish();
-            Intent intent = new Intent(NotesAct.this, WriteNotesActivity.class);
-            startActivity(intent);
-        }
-    };
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.notes_activity_menu, menu);
-        return true;
-    }
-
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.new_pass_action:
-                Intent intent = new Intent(NotesAct.this, NewPassActivity.class);
-                startActivity(intent);
-                finish();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-}
