+++ app/src/main/res/layout/activity_notes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NotesAct">
+    tools:context=".NotesListActivity">
 
     <ListView
         android:id="@+id/listView"
