+++ app/src/main/res/layout/content_notes.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".NotesAct"
+    tools:context=".NotesListActivity"
     tools:showIn="@layout/activity_notes">
 
 </android.support.constraint.ConstraintLayout>
